# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "version": "1.0.0"
    },
    "paths": {
        "/cache/checkConnection": {
            "get": {
                "tags": [
                    "Cache"
                ],
                "summary": "Check cache connection",
                "operationId": "340465e746a0d59117fae52123eb503e",
                "responses": {
                    "200": {
                        "description": "Check cache connection result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cache/flushAll": {
            "delete": {
                "tags": [
                    "Cache"
                ],
                "summary": "Flush all cache",
                "operationId": "6026c37a3e991e304e2e459bb4e6d564",
                "responses": {
                    "200": {
                        "description": "Flush all cache",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get groups",
                "operationId": "8103ba310d86370d7b29d5fd1a6c9763",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": true
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Per page",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get multiple groups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found any group(s)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get group by id",
                "operationId": "f534366b4ba7b94fefc2e37bbcb0cb13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get group by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found any group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups/{id}/topics": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get list topics by id",
                "operationId": "b039a329d8e99f576a90ee317b6aa91d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get list topics by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found any group or topic(s)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/topics/{id}": {
            "get": {
                "tags": [
                    "Topics"
                ],
                "summary": "Get topic by id",
                "operationId": "878892a7bad9c5ce48d34f34ed7202f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get topic by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ApiResult": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Cache",
            "description": "Cache"
        },
        {
            "name": "Groups",
            "description": "Groups"
        },
        {
            "name": "Topics",
            "description": "Topics"
        }
    ]
}